numbers-processor-consumer-producer-container/requirements.txt
kafka-python
inflect
